fix raner multi-view training
1. fix padding of `origin_mask` in `data_collators/base.py` and `batch.py`
2. fix tensor sizes of multi-view training in `sequence_labeling_model.py`

+++ adaseq/data/batch.py
@@ -47,7 +47,7 @@ class DataBatch(Mapping):
             elif k in self.keep_fields:
                 continue
             else:
-                batch[k] = torch.tensor(batch[k], dtype=torch.long)
+                batch[k] = torch.tensor(batch[k])
         return batch
 
     def to(self, device):
